+++ Entities/Order.cs
@@ -9,7 +9,7 @@ public partial class Order
 
     public DateTime? OderDate { get; set; }
 
-    public decimal OrderSum { get; set; }
+    public double OrderSum { get; set; }
 
     public int UserId { get; set; }
